SeatReservation Bug Fixes - Checks for seat reservation logic was slightly wrong. Added more comments so that it's clear. - Basically it was falsely throwing errors for when canceling bookings. - setOccupant check done on wrong variable. Should be check on whether booking was stored already, not if booking is set to null.

+++ wingit/src/main/java/com/G2T5203/wingit/booking/BookingController.java
@@ -68,7 +68,10 @@ public class BookingController {
     }
 
     @DeleteMapping("bookings/delete/{bookingId}")
-    public void deleteBooking(@PathVariable int bookingId) {
+    public void deleteBooking(@PathVariable int bookingId,
+                              @AuthenticationPrincipal UserDetails userDetails,
+                              @AuthenticationPrincipal Jwt jwt) {
+        checkIfNotUserNorAdmin(service.getBookingUserUsername(bookingId), userDetails, jwt);
         try {
             service.deleteBookingById(bookingId);
         } catch (BookingNotFoundException e) {
@@ -93,7 +96,10 @@ public class BookingController {
     }
 
     @PutMapping("bookings/markAsPaid/{bookingId}")
-    public void markBookingAsPaid(@PathVariable int bookingId) {
+    public void markBookingAsPaid(@PathVariable int bookingId,
+                                  @AuthenticationPrincipal UserDetails userDetails,
+                                  @AuthenticationPrincipal Jwt jwt) {
+        checkIfNotUserNorAdmin(service.getBookingUserUsername(bookingId), userDetails, jwt);
         try {
             service.markBookingAsPaid(bookingId);
         } catch (BookingNotFoundException e) {
@@ -101,5 +107,12 @@ public class BookingController {
         } catch (Exception e) {
             throw new BookingBadRequestException(e);
         }
+    }
+
+    @GetMapping("bookings/getPriceBreakdown/{bookingId}")
+    public Map<String, Object> getPriceBreakdown(@PathVariable int bookingId) {
+        // TODO: Check and verify is same user or admin
+        // TODO: TO BE IMPLEMENTED
+        return null;
     }
 }
